(split)#define BFRAMES
git-svn-id: http://svn.xvid.org/trunk@152 060a26f8-e6a2-0410-8959-474133d486e0

+++ src/bitstream/mbcoding.h
@@ -23,8 +23,8 @@ void get_inter_block(Bitstream * bs, int16_t * block);
 
 void MBCodingBVOP(const MACROBLOCK * mb, 
 				  const int16_t qcoeff[6*64],
-				  const int16_t fcode,
-				  const int16_t bcode,
+				  const int32_t fcode,
+				  const int32_t bcode,
 				  Bitstream * bs, 
 				  Statistics * pStat);
